// src/dto/registerDetails.dto.ts
export class RegisterDetailsDto {
  productName: string;
  productImages: string;
  productPrice: string;
  totalProduct: string;
  productDescription: string;
  offerStartDate: string;
  offerEndDate: string;
  shopOpenDay: string;
  shopCloseDay: string;
  country: string;
  state: string;
  city: string;
  location: string;
}